# cmd/server/openapi.yaml
openapi: "3.1.0"
info:
  title: Bastio API
  description: AI Security Gateway API
  version: "0.1.0"
  license:
    name: FSL-1.1-ALv2
    url: https://fsl.software/FSL-1.1-ALv2.template.md
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /v1/workspace/settings:
    get:
      operationId: workspaceGetSettings
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSettings"
    patch:
      operationId: workspaceUpdateSettings
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceSettingsPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSettings"
...
  /v1/workspace/assistants:
    get:
      operationId: workspaceListAssistants
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [assistants]
                properties:
                  assistants:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceAssistant"
    post:
      operationId: workspaceCreateAssistant
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceAssistant"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceAssistant"
  /v1/workspace/assistants/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: workspaceGetAssistant
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceAssistant"
    patch:
      operationId: workspaceUpdateAssistant
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceAssistantPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceAssistant"
    delete:
      operationId: workspaceArchiveAssistant
      tags: [Workspace]
      responses:
        "204": { description: Archived }
  /v1/workspace/knowledge:
    get:
      operationId: workspaceListKnowledge
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [sources]
                properties:
                  sources:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceKnowledgeSource"
    post:
      operationId: workspaceCreateKnowledge
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceKnowledgeSource"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceKnowledgeSource"
  /v1/workspace/knowledge/upload:
    post:
      operationId: workspaceUploadKnowledge
      summary: Upload a knowledge file (multipart). Async ingestion.
      description: |
        Multipart form with a `file` field. The server stores the blob,
        creates a `workspace_knowledge_sources` row with status='pending',
        enqueues a River ingestion job, and returns the row immediately.
        The dashboard polls until status flips to 'ready' or 'failed'.
      tags: [Workspace]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
      responses:
        "201":
          description: Created (ingestion in progress)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceKnowledgeSource"
  /v1/workspace/knowledge/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: workspaceGetKnowledge
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceKnowledgeSource"
    delete:
      operationId: workspaceArchiveKnowledge
      tags: [Workspace]
      responses:
        "204": { description: Archived }
...
  /v1/workspace/settings/slug:
    put:
      operationId: workspaceSetSlug
      summary: Claim or update the workspace's branded-chat slug
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug:
                  type: string
                  pattern: "^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])?$"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSettings"
        "409":
          description: Slug already claimed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /v1/workspace/domains:
    get:
      operationId: workspaceListDomains
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [domains]
                properties:
                  domains:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceDomain"
    post:
      operationId: workspaceCreateDomain
      summary: Register a custom domain (returns DNS verification token)
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain: { type: string, format: hostname }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDomain"
        "409":
          description: Domain already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /v1/workspace/domains/{id}/verify:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: workspaceVerifyDomain
      summary: Verify domain ownership via DNS TXT record
      tags: [Workspace]
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDomain"
        "422":
          description: TXT record not found or DNS lookup failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /v1/workspace/domains/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: workspaceDeleteDomain
      tags: [Workspace]
      responses:
        "204":
          description: Removed
  /v1/workspace/conversations:
    get:
      operationId: workspaceListConversations
      tags: [Workspace]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [conversations]
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceConversationListItem"
    post:
      operationId: workspaceCreateConversation
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                assistant_id: { type: string, format: uuid }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceConversation"
  /v1/workspace/conversations/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: workspaceGetConversation
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceConversation"
    patch:
      operationId: workspaceRenameConversation
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
      responses:
        "204": { description: Renamed }
    delete:
      operationId: workspaceArchiveConversation
      tags: [Workspace]
      responses:
        "204": { description: Archived }
  /v1/workspace/conversations/{id}/messages:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: workspaceListMessages
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [messages]
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceMessage"
    post:
      operationId: workspaceSendMessage
      summary: Send a message (non-streaming, returns full assistant reply)
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceSendMessageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [user_message, assistant_message]
                properties:
                  user_message:
                    $ref: "#/components/schemas/WorkspaceMessage"
                  assistant_message:
                    $ref: "#/components/schemas/WorkspaceMessage"
  /v1/workspace/conversations/{id}/messages/stream:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: workspaceStreamMessage
      summary: Send a message and stream tokens via Server-Sent Events
      description: |
        Returns a `text/event-stream` with three event types:
        * `token` — `{ "delta": "<text>" }` repeated until the response is complete
        * `done` — `{ "message": <Message> }` with the persisted assistant message
        * `error` — `{ "error": "<message>" }` on transport or provider failure
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceSendMessageRequest"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
...
  /v1/workspace/members/{userID}:
    parameters:
      - in: path
        name: userID
        required: true
        schema: { type: string }
    delete:
      operationId: workspaceRemoveMember
      tags: [Workspace]
      responses:
        "204": { description: Removed }
  /v1/workspace/invitations:
    get:
      operationId: workspaceListInvitations
      tags: [Workspace]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [invitations]
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceInvitation"
    post:
      operationId: workspaceCreateInvitation
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [owner, admin, member, viewer] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [invitation, token]
                properties:
                  invitation:
                    $ref: "#/components/schemas/WorkspaceInvitation"
                  token: { type: string }
...
  /v1/workspace/invitations/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: workspaceRevokeInvitation
      tags: [Workspace]
      responses:
        "204": { description: Revoked }
...
  /v1/workspace/analytics/daily:
    get:
      operationId: workspaceAnalyticsDaily
      tags: [Workspace]
      parameters:
        - in: query
          name: days
          schema: { type: integer, default: 14, maximum: 90 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [days]
                properties:
                  days:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkspaceDailyUsagePoint"
...
  /v1/workspace/owner/transfer:
    post:
      operationId: workspaceTransferOwnership
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_owner_user_id]
              properties:
                new_owner_user_id: { type: string }
      responses:
        "204":
          description: Owner transferred
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer sk-bastio-..."
  schemas:
...
    WorkspaceSettings:
      type: object
      required: [customer_id, branding, seat_limit, retention_days, billing_mode, allowed_models, created_at, updated_at]
      properties:
        customer_id: { type: string, format: uuid }
        branding: { type: object, additionalProperties: true }
        default_assistant_id: { type: string, format: uuid, nullable: true }
        seat_limit: { type: integer }
        retention_days: { type: integer }
        spend_cap_cents: { type: integer, nullable: true }
        billing_mode: { type: string, enum: [platform_keys, byo_keys] }
        allowed_models:
          type: array
          items: { $ref: "#/components/schemas/WorkspaceAllowedModel" }
          description: |
            Strict whitelist of LLM provider+model pairs surfaced in
            the employee chat's model picker. Empty = all curated
            defaults available; non-empty = strict whitelist (also
            enforced server-side as defense-in-depth).
        onboarding_completed_at: { type: string, format: date-time, nullable: true }
        slug: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...